[TEST] #2 Add vsms_util.py
This module is for vsms related utility functions (for AddResource, ModifyResource, etc).

Add the mod_<> functions.  These are identical the code paths in VSMS 08_modifyresource.py except for mod_net_res().

mod_net_res() - In the case of Xen, the bridge name needs to be verified.  This is because the network interface type is converted to a bridge interface by libvirt.

Updates:
 -Add generic print function

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>