[TEST] #2 Fix potiential false positive in ESD 01.
Test needs to verify the number of VSSD and RASD elemnts returned were the values expected.

Instead of calling EnumerateInstances() for the VSSD and RASD instances, this test should call getInstance() to get the expected instances.  Most of this test was written to support that.

Test now also supports XenFV.

Updates 1 to 2:
  -Uncomment cxml.undefine() on line 147.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>