[TEST] #2 Add compare_all_prop().
This function takes an instance from the list returned bu the Associators() call and an instance from a getInstance() / enumerate_instances() call.  It compares all of the property values to ensure the instances are identical.

This function could probably be added to a fair number of tests as an additional verification check.

Updates from 1 to 2:
  -Added logger import statement
  -Included FAIL in the ReturnCodes import.
  -Tested failure path to be sure error message is logged and FAIL is returned properly.


Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>