[TEST] #2 Remove guest creation from LogicalDisk - 02_nodevs.py.
The point of this test is to ensure no devices are returned from EnumerateInstances() in the case where no guest is defined (or in the case where only Domain-0 is defined).

Also add support for LXC.

In the case where devices.enumerate() encountered a CIM error, raise an exception instead of printing an error.

Updates from 1 to 2:
  -Fix spacing in clean_system()

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>