[TEST] #2 Fix potiential false positive in CS 02.
Also make it so that the test skips if domain_list() returns anything > 0.

This test used to be a Xen only test, so the test was only being run if there were no other guests besides Dom0.  However, CS EnumInstances treats Dom0 as any other guest, so it only makes sense to run this test on KVM and LXC.

If computersystem.enumerate() encounters an exception, raise an exception so that the calling test case fail appropriately.  We could return FAIL in this case, but raising an exception is more descriptive.  Also, other tests that call computersystem.enumerate() don't need to be modified because the number of arguments returned doesn't change.

Updates from 1 to 2:
  -Remove Xen and XenFV from the supported platforms list.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>