[TEST] #2 Add VSMS 10_referenced_config.py.
This test creates a guest using default RASDS.  Then another guest is defined using a reference of the first guest an the value for the ReferencedConfiguration parameter.  This second guest is created with an additional network interface.

The test verifies that the second guest is defined properly.  It also verifies that the second guest has 2 network interfaces: one with a mac address that matches that of the first guest and one with a different mac address.

This test current fails on LXC because the test suite doesn't pass a console device to the DefineSystem().  Will fix this in a follow up patch.

Updates:
  -Create a network pool and then destroy the pool when test is complete.
  -Use VSSD ref instead of CS ref

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>