[TEST] #3 Fix potiential false positive in AC 01.
This test needs to verify enum of AC returns the same number of instances as the number of instances returned by enum of MemoryPool + ProcessorPool + DiskPool + NetworkPool.

Also, cleaned up the logic to verify that the ResourceType and the InstanceIDs of the AC instances match the Pool instances they're describing.

Updates from 1 to 2:
  -Create a disk pool and network pool to ensure there is an instance for each pool type.
  -Add comment explaining the purpose of the test.

Updates from 2 to 3:
  -Change if len(pools) < 1 to if len(pools) < 4
  -Add log message if this if statement fails.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>