[TEST] #2 Enable test infrastructure to support writing to a temporary file.
Allow the option to print to a file in addition to stdout.  This will allow a test suite to read back the file after the test execution is complete.

If a previous log exists at the start of the run, it is removed and a new one is created.

Updates:
  -Add a new line at the end of each line when writing out to the log

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>