[TEST] Add new enumerate infrastructure...
CIM_CimtestClass, EnumNames, EnumInstances, and GetInstance will replace the existing CIM_MyClass, enumerate, enumerate_inst, and getInstance.  Since there are a lot of tests to change, the original infrastructure is left in place.  This should be removed once the tests are updated to use the new infrastructure.

There's a few problems with the existing infrastructure.  These replacesments aim to remove these issues:

-A new classes require an entry in enumclass.py for each virt type. These changes remove the requirement.

-A key_list is a necessary parameter to the enumerate calls.  This goes against the extrinsic API definitions - the CIM API does not include an equivallent paremeter for the enumerate calls.

-The enumerate() call currently takes a base classname, and a virt type.  Really, the just the classname (complete with prefix) should be passed in.  This way, SBLIM classes are supported in a more natrual way.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>