[TEST] Add NETNS toggle to const.py; update LXC define system bits.
Added LXC_netns_support to const.py - right now, this is hardcoded to False.  In the future, this will dynamically detect whether the kernel supports NETNS (which is needed to create containers guests with network support).

Cleanup default_vssd_rasd_str() so that class_dasd() is used to create a Disk RASD for LXC guests, as well as KVM/Xen.

On undefine/destroy of LXC guests, don't remove the file at LXC_init_path

Extend LXCXML to subclass VirtCIM - this allows cim_define() to be called for LXC guests.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>