[TEST] Improve libvirt virt support check in main.py
If /usr/bin/qemu does not exist on the system (or if libvirt cannot find it),
it is possible to define KVM guests but not run them.  Instead of returning a
error saying libvirt doesn't support the virt type, we need to return an error
that indicates a problem with libvirt was encountered.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>