[TEST] Update ETCP tests to expect instance from get_host_info().
For EC 01, make sure namespace is set as interop only after EnumInstances() for the result classes have been called.  Cleanup XFAIL logic.

Also, handle __iter__, __repr__, and items functions in enumclass.py for Cimtest
instances.  This allows for printing for Cimtest instances without encountering
an error.  Also allows the caller to call inst.items() to get a list of
instance properties - the same that can be done with instances of type
CIMInstance.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>