[TEST] Remove XFAIL from EC 01 and 02.
Fixed 01:
Update to work with CRS.

Also update the logic so that items aren't removed from the managed_ele list
when a match is properly found.  It's possible for multiple profiles to return
the same managed element, so removing the elements from this list would cause
a false failure.

Instead, maintain a found list.  If a managed element isn't in the found list,
return an error.

Fixed 02:
Fixed to work with SBLIM host instance as well as libvirt-cim host instance.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>