[TEST] #2 Remove try_getinstance() from VSMSD - 02_vsmsd_gi_errs.py
try_getinstance() is no longer needed - this function can be implemented using
functions from enumclass.py.  Plus, a conn needs to be passed to the function,
which is poor function design.

Removed the "invalid_instid_keyname" case - passing an invalid keyname only
tests the CIMOM, it does not test the providers.

Updates:
 -Change "return PASS" to "return status"

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>