[TEST] Modifying the 02_np_gi_errors.py of NetworkPort.
Modifying the 02_np_gi_errors.py of NetworkPort to be skipped for LXC when LXC_netns_support is false.
As of now LXC_netns_support is hardcoded to be false in const.py.
We need to update the tc once we have an appropriate function which verifies the NETNS support in the kernel.

Signed-off-by: Deepti B. Kalakeri <deeptik@linux.vnet.ibm.com>