[TEST] #3 Update VSMS 04_definesystem_ers.py to use cim_define()
Add a err_rc and err_desc elements to the VirtCIM class.  This will allow
us to check the error code and error description in case of failed CIM
method calls.

Add two functions: set_sys_settings() and set_res_settings().  These allow the
caller to set the VSSD and RASDs after the initial init() of the VirtCIM class.

Some of the logic in cim_define() was changed to allow the passing of just
a single RASD to DefineSystem() (instead of passing all the RASDs).

Updates from 2 to 3:
  -Rebased on current tree

Updates from 1 to 2:
  -Create function to verify return code and error description if define fails

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>