[TEST] #3 Fix verify_rasds() and verify_devices() in SD 02_reverse.py
Since VSSDC should return RASDs that all correspond to the guest, there's
no need to parse the InstanceID to see if the guest name matches the guest
we've created.

Also, fix the logic of the verify functions to be a little cleaner. Compare
all the properties of the instances, not just the InstanceIDs/DeviceIDs.

Remove parse_instance_id() from verify_devices(); guest and dev_id were not
being used in the function.

Updates from 2 to 3:
  -Fix typo in setup_env().  cmxl should be cxml
  -Fixed bug in cim_define() that was causing this to fail with LXC

Updates from 1 to 2:
  -Removed unnecessary undefine() call
  -Moved the verification of the number of RASDs before the for loop - in the
   case of LXC, we remove one of the RASDs from this list so we don't verify it,
   but we still want to verify the appropriate number of RASDs was returned.
  -Print an error message if comparing all the instance properties fails.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>