[TEST] Fixing the EAFP/01_forward.py and VSSDC/02_reverse.py by using cim_define() and cim_start().
Common changes to EAFP/01_forward.py and VSSDC/02_reverse.py:
-------------------------------------------------------------
Modified the tc to use cim_define() and cim_start() instead of create()
which was causing the tc to fail with Xen.

Changes to EAFP/01_forward.py:
------------------------------
1) Added try, except block.

Changes to VSSDC/02_reverse.py:
-------------------------------
1) Removed unnecessary Global import statement.
2) Did some Indentation changes to the tc to fit 80 columns.

The test cases can be further improved in the next cycle to make use of EnumInstances() and compare_all_prop()
to reduce the code and make it look more cleaner.

Tested for Xen with current sources.

Signed-off-by: Deepti B. Kalakeri <deeptik@linux.vnet.ibm.com>