[TEST] #4 Make sure network pool is created with a random IP.
This will help prevent overlap with existing networks.  However, this won't
completely prevent the issue.  We'd need to do test to see if the IP address
is already used by a different network. If the network is in use, then another
IP address should be used.

Update from 3 to 4:
   -Remove changes made for debugging purposes:
     *Uisng a 192.168.122 value for the subnet instead of 192.168.123
     *Overriding addr value

Updates from 2 to 3:
   -If IP address is already in use, return from the _init_() call.

Updates from 1 to 2:
  -Check to see if pool with the specified IP range exists, if so, return a
   meaningful error
  -Since the NetXML generates a random IP, creating a NetXML object in the
   destroy call will generate an object with a different IP.  Instead, use
   the virsh call to destroy the network.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>