[TEST] Fixing the run() fn of vxml.py.
Fixing run() fn of vxml.py to be able to run virsh commands on the remote machine.
We need this for successful execution of the ssh remote commands like net-createa via virsh,
otherwise the remote execution of the command fails when the file is not locally present on the remote machine.

Signed-off-by: Deepti B. Kalakeri <deeptik@linux.vnet.ibm.com>