[TEST] Remove default emulator in VSSD, also fix path of default in const.py
libvirt will determine the proper emulator to use based on the capabilities
of the system.  No need to define a default.  If a different emulator is needed
(i.e. because qemu / kvm is installed in a non-standard location), the test
case can specify an emulator using the Emulator attribute of the VSSD.


Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>