[TEST] #2 Move cleanup_guest_netpool() to vsmigrations.py
Migration test 06 - 08 duplicate this same code.

Also, if the migration is a localhost one, set the hostname to localhost.
Otherwise, the providers will return an error saying the guest already exists
on the target (because the providers haven't detected a localhost migration).

If the target system name is localhost, the migration will always be a local
migration.  Be sure to set remote_migration accordingly.

Update from 1 to 2:
  -Change destroy() to cim_destroy()

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>