[TEST] Fix rasd_init_list() - XenFV guests with newer vesions of libvirt...
Use ps2 input devices, not usb.  Set ps2 as the default, otherwise, the guest
could end up with both ps2 and usb devices, which breaks some of the tests.
This is because the test is only expecting one input device, but libvirt adds
an additional ps2 device (in the case usb is specified).

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>