[TEST] Remove stale XenFV guest if migration fails in VSMS/02
Sometimes xend gets hung doing a local migration of XenFV guests.  The guest
will show up in "xm list" as "migrating-guest_name". The problem with this is
that libvirt sees the guest as "guest_name", so libvirt is unable to
undefine the guest.  In this case, we need to do an xm destroy to remove the
guest.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>