[TEST] Fix VSMS to do a proper check of ref config, also remove test_xml import
This test was originally designed to do the following:

1) Create a guest with a MAC interface
2) Create a second guest based on the first guest - second guest has an
   additional MAC defined.  Pass a reference to the first guest during the
   DefineSystem()
3) Verify the second guest was created with two MACs - one that is identical to
   the first guest and one that is different

The providers no longer allow a guest to have the same MAC as an existing guest.
Each MAC needs to be unique.  Therefore, this test needs to use a different
setting - disk source works for this.

Also, remove the dependency on test_xml.py - that module is not obsolete.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>