On Fedora systems default to using 'em1' instead of 'eth1'
Fedora changed the default device naming scheme, see
  http://fedoraproject.org/wiki/Features/ConsistentNetworkDeviceNaming

So if we're running on Fedora, let's "default" to "em1" although that
doesn't guarantee that we will find what we're looking for.

This code probably needs some mechanism to detect with interfaces are
available, then query those interfaces to find one that supports the
feature we want.