Add and utilize virsh_version_cmp
The string version comparisons failed when determining if "0.10.2" was
a greater version than "0.4.1" (and other similar checks). Needed to create
a version comparison method that did the right thing.

Also, not change to 12_create_netfs_storagevolume_errs.py to use "or" logic
rather than "and" logic on comparison. Other tests within the family use
"or" logic.