Add default network card name existence check
Default network card name was set as 'eth1' or 'em1'. But when os has no network card named like that, FAIL
would occured. This patch added check if default name exists in the network card list and if not, change
the default value into the 1st one in the list.

Signed-off-by: Xu Wang <cngesaint@gmail.com>