cimtest: Fix bogus RASD counting in RASD/03_rasd_errs
The number of expected RASDs for the test domain was incorrectly
computed by enumerating all RASDs on the hypervisor which will
always fail if the test domain doesn't have every possible
RASD associated.
Fixed by counting the resource settings in the virtual server
instance and comparing that against the RASDs actually found.

Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>