cimtest: Use lists instead of dictionaries
InputRASD's and an uncoming ControllerRASD object can appear multiple
times in the guest XML.  Currently the code uses dictionaries with the
ClassName of the object as the key to the dictionary.  We cannot do that
as we move forward, so convert to using a list.  The list will be a tuple
of the ClassName and the object previously stored in the dictionary key.

InputRASD's currently only have a mouse or tablet, but as of libvirt v1.2.2
there will be a keyboard device added.

ControllerRASD's will be added shortly and there are usually 2 or 3 per
domain.

Signed-off-by: John Ferlan <jferlan@redhat.com>