cimtest: VSSDC - 02-reverse.py - adjust iteration of association
Upcoming changes to libvirt-cim allow for multiple types of objects
per class name for the Input class and a new upcoming Controller class.

Adjust the logic in the code to change the iteration to handle the
situation where there could be more than one type of device in the class.

An innocent bystander of this change was the extra check in the code:

        if status == FAIL and virt == "LXC":
            return XFAIL_RC(bug_libvirt)

Since I wasn't sure what this was testing and perhaps it's actually
been fixed - I removed the check. If it's still a bug, then it needs
to be fixed and there's only one way to find out!